# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: IronClaw Control-Plane API
  version: 0.1.0
  summary: Gateway change-approval, audit, and registry administration for a self-hosted IronClaw control-plane.
  description: |
    The control-plane exposes three surfaces:

    - **Gateway** (`/v1/changes`, `/v1/audit`) — submit capability changes, list
      what is pending, record human approve/reject decisions, and read the
      append-only audit log. Every control-plane mutation (persona, enabled
      tools, packages, wiring, permissions, mounts, create_agent) flows through
      here; there is no bypass path.
    - **Registry** (`/v1/registry/...`) — administer agent groups, messaging
      groups, wirings, users, roles, members, destinations, and read sessions.
      These endpoints return `503` until a registry backend is attached.
    - **Operational probes** (`/healthz`, `/readyz`, `/metrics`).

    All `POST`/`PUT` bodies are JSON. The Go JSON decoder is case-insensitive on
    keys; the property names below are the canonical forms the server **emits**.
  license:
    name: MIT
    url: https://github.com/nivardsec/ironclaw/blob/main/LICENSE
  contact:
    name: IronClaw
    url: https://github.com/nivardsec/ironclaw
...
servers:
  - url: http://127.0.0.1:8787
    description: Local development (`--dev`, loopback bind).
  - url: http://{tailnetHost}:{port}
    description: Production — bound to the control-plane's Tailscale interface address (no public port).
    variables:
...
      port:
...

...
security:
  - bearerAuth: []
...
paths:
...
  /v1/audit:
    get:
      tags: [gateway]
      summary: Read recent audit entries.
      description: Returns an empty list when no audit path is attached. Entries are append-only JSONL records.
      operationId: readAudit
      parameters:
        - name: limit
          in: query
          required: false
          description: Maximum entries to return (most recent first). Defaults to 100; non-positive values are ignored.
          schema:
            type: integer
            minimum: 1
            default: 100
      responses:
        "200":
          description: Recent audit entries (possibly empty).
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditEntry"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/registry/agent-groups/{id}:
    put:
      tags: [registry]
      summary: Create or update an agent group.
      description: The path `id` is authoritative and overrides any `ID` in the body.
      operationId: putAgentGroup
      parameters:
        - $ref: "#/components/parameters/IDPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentGroup"
      responses:
        "200":
          description: The stored agent group.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentGroup"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/RegistryUnavailable"
    get:
      tags: [registry]
      summary: Fetch an agent group by id.
      operationId: getAgentGroup
      parameters:
        - $ref: "#/components/parameters/IDPath"
      responses:
        "200":
          description: The agent group.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentGroup"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/RegistryUnavailable"
...
  /v1/registry/messaging-groups/{id}:
    get:
      tags: [registry]
      summary: Fetch a messaging group by id.
      operationId: getMessagingGroup
      parameters:
        - $ref: "#/components/parameters/IDPath"
      responses:
        "200":
          description: The messaging group.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessagingGroup"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/RegistryUnavailable"
  /v1/registry/messaging-groups/{id}/wirings:
    get:
      tags: [registry]
      summary: List the wirings for a messaging group.
      operationId: listWirings
      parameters:
        - $ref: "#/components/parameters/IDPath"
      responses:
        "200":
          description: The wirings (possibly empty).
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Wiring"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/RegistryUnavailable"
...
  /v1/registry/members:
    post:
      tags: [registry]
      summary: Add a user as a member of an agent group.
      operationId: addMember
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Member"
      responses:
        "201":
          description: The added member.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/RegistryUnavailable"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Static bearer token, supplied to the server via `IRONCLAW_API_TOKEN`
        (or `--token` for `ironctl`). It is **all-or-nothing** — there are no
        per-endpoint scopes; a valid token authorizes every non-probe endpoint.
        The token is compared in constant time. `/healthz` and `/readyz` are
        always exempt. When the server is started without a token the check is
        disabled and the mesh (Tailscale) boundary is the sole control.
  parameters:
    IDPath:
      name: id
      in: path
      required: true
      description: The resource id (authoritative for create/update).
      schema:
        type: string
...
  responses:
    Unauthorized:
      description: Missing or invalid bearer token (only when the server is started with a token).
      content:
        text/plain:
          schema:
            type: string
            example: unauthorized
    BadRequest:
      description: Malformed JSON body or invalid field values.
      content:
        text/plain:
          schema:
            type: string
    NotFound:
      description: Resource not found.
      content:
        text/plain:
          schema:
            type: string
    InternalError:
      description: Internal error reading from the backing store.
      content:
        text/plain:
          schema:
            type: string
...
    RegistryUnavailable:
      description: No registry backend is attached to this control-plane.
      content:
        text/plain:
          schema:
            type: string
            example: registry not configured
...
  schemas:
...
    ChangeID:
      type: string
      description: Gateway change id (server-assigned `chg_<timestamp>` when omitted).
      examples: ["chg_20260617120000.000000000"]
    UserID:
      type: string
      description: Platform identity, formatted `<channel>:<handle>`.
      examples: ["slack:U123", "telegram:42"]
    AgentGroupID:
      type: string
    MessagingGroupID:
      type: string
...
    ChangeKind:
      type: string
      description: The control-plane mutation kind.
      enum:
        - persona
        - enabled_tools
        - packages
        - wiring
        - permissions
        - mounts
        - create_agent
    UnknownSenderPolicy:
      type: string
      enum: [strict, request_approval, public]
      default: strict
    EngageMode:
      type: string
      enum: [pattern, mention, mention-sticky]
    SenderScope:
      type: string
      enum: [all, known]
    IgnoredMessagePolicy:
      type: string
      enum: [drop, accumulate]
    SessionMode:
      type: string
      enum: [shared, per-thread, agent-shared]
...
    AuditStage:
      type: string
      enum: [submit, verdict, decision, apply]

    # --- gateway types (internal/contract/gateway.go, internal/host/gateway) ---
...
    AuditEntry:
      type: object
      description: One append-only audit record (emitted as JSONL on disk).
      properties:
        time:
          type: string
          format: date-time
        stage:
          $ref: "#/components/schemas/AuditStage"
        changeId:
          $ref: "#/components/schemas/ChangeID"
        kind:
          $ref: "#/components/schemas/ChangeKind"
        detail:
          type: string
      required: [time, stage, changeId]

    # --- registry types (internal/host/registry/registry.go; no JSON tags → capitalized keys) ---
    AgentGroup:
      type: object
      description: A configured agent (workspace, persona, container config). Only host-facing identity fields are exposed here.
      properties:
        ID:
          $ref: "#/components/schemas/AgentGroupID"
        Name:
          type: string
        Folder:
          type: string
...
    MessagingGroup:
      type: object
      description: A single chat/channel on one platform, identified by the (ChannelType, PlatformID, Instance) triple.
      properties:
        ID:
          $ref: "#/components/schemas/MessagingGroupID"
        ChannelType:
          type: string
        PlatformID:
          type: string
        Instance:
          type: string
        IsGroup:
          type: boolean
        UnknownSenderPolicy:
          $ref: "#/components/schemas/UnknownSenderPolicy"
...
    Wiring:
      type: object
      description: Links a messaging group to an agent group with engage/session policy.
      properties:
        ID:
          type: string
        MessagingGroupID:
          $ref: "#/components/schemas/MessagingGroupID"
        AgentGroupID:
          $ref: "#/components/schemas/AgentGroupID"
        EngageMode:
          $ref: "#/components/schemas/EngageMode"
        EngagePattern:
          type: string
        SenderScope:
          $ref: "#/components/schemas/SenderScope"
        IgnoredMessagePolicy:
          $ref: "#/components/schemas/IgnoredMessagePolicy"
        SessionMode:
          $ref: "#/components/schemas/SessionMode"
        Priority:
          type: integer
...
    Member:
      type: object
      description: Records unprivileged access of a user to an agent group.
      properties:
        UserID:
          $ref: "#/components/schemas/UserID"
        AgentGroupID:
          $ref: "#/components/schemas/AgentGroupID"
      required: [UserID, AgentGroupID]
...